Made compressedMrefIndexToIndex() return an optional to account for failing index mapping.
It is unclear, why the mapping should fail. The optional guarantees that the calling code has to handle the situation. There are many places where the calling code implements a default action, and other placess (e.g. AnnotationProcessor.cpp:227) where an error should occur (used to be segmentation fault, now is exception).

Signed-off-by: Philip R. Kensche <p.kensche@dkfz.de>